pyproject.toml
[tool.poetry]
name = "api development on flask"
version = "0.1.0"
description = ""
authors = ["Admin <Dexter2366@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.3"
flask-sqlalchemy = "^3.1.1"
sqlalchemy = "^2.0.21"
marshmallow = "^3.20.1"
flask-restx = "^1.1.0"
pyjwt = "^2.8.0"
pytest = "^7.4.3"
flask-migrate = "^4.0.5"
flask-script = "^2.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
